Refactored the delete function.

+++ doubly_linked_list/doubly_linked_list.py
@@ -89,6 +89,8 @@ class DoublyLinkedList:
     current head's next node the new head of the List.
     Returns the value of the removed Node."""
     def remove_from_head(self):
+##### Code From Lecture Review #####
+
         self.length -= 1
         if not self.head:
             return None
@@ -106,6 +108,8 @@ class DoublyLinkedList:
     as the new tail of the list. Don't forget to handle 
     the old tail node's next pointer accordingly."""
     def add_to_tail(self, value):
+##### Code From Lecture Review #####
+
         self.length += 1
         new_list_node = ListNode(value)
         if self.head is None and self.tail is None:
@@ -124,6 +128,8 @@ class DoublyLinkedList:
     current tail's previous node the new tail of the List.
     Returns the value of the removed Node."""
     def remove_from_tail(self):
+##### Code From Lecture Review #####
+
         self.length -= 1
         if not self.tail:
             return None
@@ -140,6 +146,8 @@ class DoublyLinkedList:
     """Removes the input node from its current spot in the 
     List and inserts it as the new head node of the List."""
     def move_to_front(self, node):
+##### Code From Lecture Review #####
+
         # self.length -= 1
         if node is self.head:
             pass
@@ -156,6 +164,8 @@ class DoublyLinkedList:
     """Removes the input node from its current spot in the 
     List and inserts it as the new tail node of the List."""
     def move_to_end(self, node):
+##### Code From Lecture Review #####
+
         # self.length -= 1
         if node is self.tail:
             pass
@@ -172,14 +182,35 @@ class DoublyLinkedList:
     """Removes a node from the list and handles cases where
     the node was the head or the tail"""
     def delete(self, node):
-        # if node.next is None:
-        #     self.remove_from_tail(node)
-        # elif node.prev is None:
-        #     self.remove_from_head(node)
-        node.delete()
+##### Code From Lecture Review #####
+        if self.head is None and self.tail is None:
+            return
+        self.length -= 1
+        if self.head == self.tail and node == self.head:
+            self.head == None
+            self.tail == None
+        if self.head == node:
+            self.head = node.next
+            node.delete()
+        if self.tail == node:
+            self.tail = node.prev
+            node.delete()
+        else:
+            node.delete()
+
+
+##### My Original Code (unfinished) #####
+
+        # # if node.next is None:
+        # #     self.remove_from_tail(node)
+        # # elif node.prev is None:
+        # #     self.remove_from_head(node)
+        # node.delete()
         
     """Returns the highest value currently in the list"""
     def get_max(self):
+##### Code From Lecture Review #####
+
         if self.head:
             node = self.head
             current_max = node.value
